requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-JWT-UI==3.36.0
python-dotenv==0.19.2
PyJWT==2.30
Werkzeug==2.1.1
pytest==7.1.1
pyodbc==4.0.32
Flask-Injector==0.15.0
Flask-Swagger-UI==4.11.1